[TASK] Move page_backend_layout to suggest
The package `page_backend_layout` is not
required for the basic functionality of
this extension.

To get a clear state to all other academic
extensions, remove `page_backend_layout`
from require and moved to suggest. On top
of that, conflict for this package is added
for unwanted versions to be used with this
extension version.