[TASK] Remove version from `composer.json`
Setting version within a `composer.json` of
a package is a real bad habit and **must**
be avoided at all costs.

Even for the only suitable usage as project
root composer.json, there is a way around
this using the `COMPOSER_ROOT_VERSION` env
variable and also a prefered alternative
over the version key, rendering it literally
unusable.

Remove it.